Updated Screen5 to import Screen3 instead of Screen1
This commit updates the import statement in the Screen5 class from `telas.screen1` to `telas.screen3`. It also removes the `vizinhanca_vertice` method and related code that was used to identify the neighbors of a vertex in a graph. Instead, the method of identifying neighbors is assumed to be implemented in the Screen3 class.